+++ apps/test-server/src/reality/issue.tsx
@@ -5,7 +5,6 @@ import {
   enableDebugTool,
   Entity,
   EntityRef,
-  getSession,
   Model,
   ModelAsset,
   ModelEntity,