+++ packages/core/src/platform-adapter/android/AndroidPlatform.ts
@@ -40,7 +40,7 @@ export class AndroidPlatform implements PlatformAbility {
             resolve(CommandResultSuccess(result.data))
           } else {
             const { code, message } = result.data as JSBError
-            reject(CommandResultFailure(code, message))
+            resolve(CommandResultFailure(code, message))
           }
         })
 
@@ -61,7 +61,7 @@ export class AndroidPlatform implements PlatformAbility {
           `AndroidPlatform cmd: ${cmd}, msg: ${msg} error: ${error}`,
         )
         const { code, message } = error as JSBError
-        reject(CommandResultFailure(code, message))
+        resolve(CommandResultFailure(code, message))
       }
     })
   }
@@ -91,10 +91,10 @@ export class AndroidPlatform implements PlatformAbility {
     // Make the page renderable through window.open
     windowProxy?.open('about:blank', '_self')
     // Polling to check if SpatialId injection is successful
-    while (!windowProxy?.SpatialId) {
+    while (!windowProxy?.__SpatialId) {
       await new Promise(resolve => setTimeout(resolve, 16))
     }
-    let spatialId = windowProxy?.SpatialId
+    let spatialId = windowProxy?.__SpatialId
     creatingElementCount--
     return Promise.resolve(
       CommandResultSuccess({ windowProxy: windowProxy, id: spatialId }),