+++ packages/core/src/platform-adapter/index.ts
@@ -1,14 +1,24 @@
-import { VisionOSPlatform } from './vision-os/VisionOSPlatform'
+import { isSSREnv } from '../ssr-polyfill'
 import { PlatformAbility } from './interface'
-import { AndroidPlatform } from './android/AndroidPlatform'
+import { SSRPlatform } from './ssr/SSRPlatform'
 import { PuppeteerPlatform } from './puppeteer/PuppeteerPlatform'
 
 export function createPlatform(): PlatformAbility {
-  console.log('window UserAgent:', window.navigator.userAgent)
-  if (window.navigator.userAgent.includes('Android')) {
+  if (isSSREnv()) {
+    return new SSRPlatform()
+  }
+
+  if (
+    window.navigator.userAgent.includes('Android') ||
+    window.navigator.userAgent.includes('Linux')
+  ) {
+    const AndroidPlatform = require('./android/AndroidPlatform').AndroidPlatform
     return new AndroidPlatform()
   } else if (window.navigator.userAgent.includes('Puppeteer')) {
     return new PuppeteerPlatform()
+  } else {
+    const VisionOSPlatform =
+      require('./vision-os/VisionOSPlatform').VisionOSPlatform
+    return new VisionOSPlatform()
   }
-  return new VisionOSPlatform()
 }