+++ packages/androidXR/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @webspatial/platform-androidxrapp
 
+## 1.0.5
+
+## 1.0.4
+
 ## 1.0.3
 
 ## 1.0.2