+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @webspatial/builder
 
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies [d8cc711]
+  - @webspatial/platform-visionos@1.0.5
+
+## 1.0.4
+
+### Patch Changes
+
+- @webspatial/platform-visionos@1.0.4
+
 ## 1.0.3
 
 ### Patch Changes