+++ tests/ci-test/CHANGELOG.md
@@ -1,5 +1,22 @@
 # ci-test
 
+## 0.0.16
+
+### Patch Changes
+
+- @webspatial/builder@1.0.5
+- @webspatial/core-sdk@1.0.5
+- @webspatial/react-sdk@1.0.5
+
+## 0.0.15
+
+### Patch Changes
+
+- Updated dependencies [67a908c]
+  - @webspatial/react-sdk@1.0.4
+  - @webspatial/core-sdk@1.0.4
+  - @webspatial/builder@1.0.4
+
 ## 0.0.14
 
 ### Patch Changes